intitial commit
this impliments the blacklist for #1691, though it does pull a function from the educator decorator file, which is a bit messy. All that needs to be done is drop a AI_BANNER_STUDENT_BLACKLIST in the env variables with a list following of the (student) memberships you want blacklisted from the banner ( roxana proveded the following: 2, 3, 87, 58 ). Students in these memberships will not see the ai_banner until the membership is removed from the list.