Student AI banner blacklist (#1697)
* intitial commit

this impliments the blacklist for #1691, though it does pull a function from the educator decorator file, which is a bit messy. All that needs to be done is drop a AI_BANNER_STUDENT_BLACKLIST in the env variables with a list following of the (student) memberships you want blacklisted from the banner ( roxana proveded the following: 2, 3, 87, 58 ). Students in these memberships will not see the ai_banner until the membership is removed from the list.

* updated as per pr comments

* updates as per PR comments

if there is a 'default blank string' and an 'if id else' then there is a int conversion error on the black list when there is no env variable defined, so i can't use the super short version.

* removed 'None'