Fix for the issue plus additional debugging #1852
The target was sometimes the button gliph rather than the button; the addition of .closest( id = visual media ) ensures that we always propogate up through children till we get to our intended target.

I also added more error logging, both for the case of an unexpected exception and us not finding the intended target (the original issue).